Moved filters to ext and fixed some imports
