+++ telegram/ext/_handlers/callbackqueryhandler.py
@@ -24,7 +24,7 @@ from typing import TYPE_CHECKING, Any, Callable, Match, Optional, Pattern, TypeV
 from telegram import Update
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._handlers.basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, HandlerCallback
 
 if TYPE_CHECKING:
