+++ telegram/ext/_handlers/choseninlineresulthandler.py
@@ -16,29 +16,29 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the ChatMemberHandler class."""
-from typing import Final, Optional, TypeVar
+"""This module contains the ChosenInlineResultHandler class."""
+import re
+from typing import TYPE_CHECKING, Any, Match, Optional, Pattern, TypeVar, Union, cast
 
 from telegram import Update
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._handlers.basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, HandlerCallback
 
 RT = TypeVar("RT")
 
+if TYPE_CHECKING:
+    from telegram.ext import Application
 
-class ChatMemberHandler(BaseHandler[Update, CCT]):
-    """Handler class to handle Telegram updates that contain a chat member update.
+
+class ChosenInlineResultHandler(BaseHandler[Update, CCT]):
+    """Handler class to handle Telegram updates that contain
+    :attr:`telegram.Update.chosen_inline_result`.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
-
-    Examples:
-        :any:`Chat Member Bot <examples.chatmemberbot>`
-
-    .. versionadded:: 13.4
 
     Args:
         callback (:term:`coroutine function`): The callback function for this handler. Will be
@@ -49,62 +49,73 @@ class ChatMemberHandler(BaseHandler[Update, CCT]):
 
             The return value of the callback is usually ignored except for the special case of
             :class:`telegram.ext.ConversationHandler`.
-        chat_member_types (:obj:`int`, optional): Pass one of :attr:`MY_CHAT_MEMBER`,
-            :attr:`CHAT_MEMBER` or :attr:`ANY_CHAT_MEMBER` to specify if this handler should handle
-            only updates with :attr:`telegram.Update.my_chat_member`,
-            :attr:`telegram.Update.chat_member` or both. Defaults to :attr:`MY_CHAT_MEMBER`.
         block (:obj:`bool`, optional): Determines whether the return value of the callback should
             be awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`. Defaults to :obj:`True`.
 
             .. seealso:: :wiki:`Concurrency`
+        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`, optional): Regex pattern. If not
+            :obj:`None`, :func:`re.match`
+            is used on :attr:`telegram.ChosenInlineResult.result_id` to determine if an update
+            should be handled by this handler. This is accessible in the callback as
+            :attr:`telegram.ext.CallbackContext.matches`.
 
+            .. versionadded:: 13.6
     Attributes:
         callback (:term:`coroutine function`): The callback function for this handler.
-        chat_member_types (:obj:`int`): Optional. Specifies if this handler should handle
-            only updates with :attr:`telegram.Update.my_chat_member`,
-            :attr:`telegram.Update.chat_member` or both.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`.
+        pattern (`Pattern`): Optional. Regex pattern to test
+            :attr:`telegram.ChosenInlineResult.result_id` against.
+
+            .. versionadded:: 13.6
 
     """
 
-    __slots__ = ("chat_member_types",)
-    MY_CHAT_MEMBER: Final[int] = -1
-    """:obj:`int`: Used as a constant to handle only :attr:`telegram.Update.my_chat_member`."""
-    CHAT_MEMBER: Final[int] = 0
-    """:obj:`int`: Used as a constant to handle only :attr:`telegram.Update.chat_member`."""
-    ANY_CHAT_MEMBER: Final[int] = 1
-    """:obj:`int`: Used as a constant to handle both :attr:`telegram.Update.my_chat_member`
-    and :attr:`telegram.Update.chat_member`."""
+    __slots__ = ("pattern",)
 
     def __init__(
         self,
         callback: HandlerCallback[Update, CCT, RT],
-        chat_member_types: int = MY_CHAT_MEMBER,
         block: DVType[bool] = DEFAULT_TRUE,
+        pattern: Optional[Union[str, Pattern[str]]] = None,
     ):
         super().__init__(callback, block=block)
 
-        self.chat_member_types: Optional[int] = chat_member_types
+        if isinstance(pattern, str):
+            pattern = re.compile(pattern)
 
-    def check_update(self, update: object) -> bool:
+        self.pattern: Optional[Union[str, Pattern[str]]] = pattern
+
+    def check_update(self, update: object) -> Optional[Union[bool, object]]:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
             update (:class:`telegram.Update` | :obj:`object`): Incoming update.
 
         Returns:
-            :obj:`bool`
+            :obj:`bool` | :obj:`re.match`
 
         """
-        if isinstance(update, Update):
-            if not (update.my_chat_member or update.chat_member):
-                return False
-            if self.chat_member_types == self.ANY_CHAT_MEMBER:
+        if isinstance(update, Update) and update.chosen_inline_result:
+            if self.pattern:
+                if match := re.match(self.pattern, update.chosen_inline_result.result_id):
+                    return match
+            else:
                 return True
-            if self.chat_member_types == self.CHAT_MEMBER:
-                return bool(update.chat_member)
-            return bool(update.my_chat_member)
-        return False
+        return None
+
+    def collect_additional_context(
+        self,
+        context: CCT,
+        update: Update,  # skipcq: BAN-B301
+        application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
+        check_result: Union[bool, Match[str]],
+    ) -> None:
+        """This function adds the matched regex pattern result to
+        :attr:`telegram.ext.CallbackContext.matches`.
+        """
+        if self.pattern:
+            check_result = cast(Match, check_result)
+            context.matches = [check_result]
