+++ telegram/ext/_handlers/messagehandler.py
@@ -16,35 +16,38 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the StringCommandHandler class."""
+"""This module contains the MessageHandler class."""
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, TypeVar, Union
 
-from typing import TYPE_CHECKING, Any, List, Optional
-
+from telegram import Update
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handlers._basehandler import BaseHandler
-from telegram.ext._utils.types import CCT, RT, HandlerCallback
+from telegram.ext import filters as filters_module
+from telegram.ext._handlers.basehandler import BaseHandler
+from telegram.ext._utils.types import CCT, HandlerCallback
 
 if TYPE_CHECKING:
     from telegram.ext import Application
 
+RT = TypeVar("RT")
 
-class StringCommandHandler(BaseHandler[str, CCT]):
-    """Handler class to handle string commands. Commands are string updates that start with
-    ``/``. The handler will add a :obj:`list` to the
-    :class:`CallbackContext` named :attr:`CallbackContext.args`. It will contain a list of strings,
-    which is the text following the command split on single whitespace characters.
 
-    Note:
-        This handler is not used to handle Telegram :class:`telegram.Update`, but strings manually
-        put in the queue. For example to send messages with the bot using command line or API.
+class MessageHandler(BaseHandler[Update, CCT]):
+    """Handler class to handle Telegram messages. They might contain text, media or status
+    updates.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Args:
-        command (:obj:`str`): The command this handler should listen for.
+        filters (:class:`telegram.ext.filters.BaseFilter`): A filter inheriting from
+            :class:`telegram.ext.filters.BaseFilter`. Standard filters can be found in
+            :mod:`telegram.ext.filters`. Filters can be combined using bitwise
+            operators (& for and, | for or, ~ for not). Passing :obj:`None` is a shortcut
+            to passing :class:`telegram.ext.filters.ALL`.
+
+            .. seealso:: :wiki:`Advanced Filters <Extensions---Advanced-Filters>`
         callback (:term:`coroutine function`): The callback function for this handler. Will be
             called when :meth:`check_update` has determined that an update should be processed by
             this handler. Callback signature::
@@ -60,7 +63,8 @@ class StringCommandHandler(BaseHandler[str, CCT]):
             .. seealso:: :wiki:`Concurrency`
 
     Attributes:
-        command (:obj:`str`): The command this handler should listen for.
+        filters (:class:`telegram.ext.filters.BaseFilter`): Only allow updates with these Filters.
+            See :mod:`telegram.ext.filters` for a full list of all available filters.
         callback (:term:`coroutine function`): The callback function for this handler.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
@@ -68,41 +72,40 @@ class StringCommandHandler(BaseHandler[str, CCT]):
 
     """
 
-    __slots__ = ("command",)
+    __slots__ = ("filters",)
 
     def __init__(
         self,
-        command: str,
-        callback: HandlerCallback[str, CCT, RT],
+        filters: Optional[filters_module.BaseFilter],
+        callback: HandlerCallback[Update, CCT, RT],
         block: DVType[bool] = DEFAULT_TRUE,
     ):
         super().__init__(callback, block=block)
-        self.command: str = command
+        self.filters: filters_module.BaseFilter = (
+            filters if filters is not None else filters_module.ALL
+        )
 
-    def check_update(self, update: object) -> Optional[List[str]]:
+    def check_update(self, update: object) -> Optional[Union[bool, Dict[str, List[Any]]]]:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
-            update (:obj:`object`): The incoming update.
+            update (:class:`telegram.Update` | :obj:`object`): Incoming update.
 
         Returns:
-            List[:obj:`str`]: List containing the text command split on whitespace.
+            :obj:`bool`
 
         """
-        if isinstance(update, str) and update.startswith("/"):
-            args = update[1:].split(" ")
-            if args[0] == self.command:
-                return args[1:]
+        if isinstance(update, Update):
+            return self.filters.check_update(update) or False
         return None
 
     def collect_additional_context(
         self,
         context: CCT,
-        update: str,  # skipcq: BAN-B301
+        update: Update,  # skipcq: BAN-B301
         application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
-        check_result: Optional[List[str]],
+        check_result: Optional[Union[bool, Dict[str, object]]],
     ) -> None:
-        """Add text after the command to :attr:`CallbackContext.args` as list, split on single
-        whitespaces.
-        """
-        context.args = check_result
+        """Adds possible output of data filters to the :class:`CallbackContext`."""
+        if isinstance(check_result, dict):
+            context.update(check_result)
