+++ telegram/ext/_handlers/shippingqueryhandler.py
@@ -14,26 +14,25 @@
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Lesser Public License for more details.
 #
-#  You should have received a copy of the GNU Lesser Public License
-#  along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the PollAnswerHandler class."""
+# You should have received a copy of the GNU Lesser Public License
+# along with this program.  If not, see [http://www.gnu.org/licenses/].
+"""This module contains the ShippingQueryHandler class."""
 
 
 from telegram import Update
-from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._handlers.basehandler import BaseHandler
 from telegram.ext._utils.types import CCT
 
 
-class PollAnswerHandler(BaseHandler[Update, CCT]):
-    """Handler class to handle Telegram updates that contain a
-    :attr:`poll answer <telegram.Update.poll_answer>`.
+class ShippingQueryHandler(BaseHandler[Update, CCT]):
+    """Handler class to handle Telegram :attr:`telegram.Update.shipping_query`.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Examples:
-        :any:`Poll Bot <examples.pollbot>`
+        :any:`Payment Bot <examples.paymentbot>`
 
     Args:
         callback (:term:`coroutine function`): The callback function for this handler. Will be
@@ -68,4 +67,4 @@ class PollAnswerHandler(BaseHandler[Update, CCT]):
             :obj:`bool`
 
         """
-        return isinstance(update, Update) and bool(update.poll_answer)
+        return isinstance(update, Update) and bool(update.shipping_query)
