+++ telegram/ext/_handlers/stringcommandhandler.py
@@ -16,27 +16,24 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the StringRegexHandler class."""
+"""This module contains the StringCommandHandler class."""
 
-import re
-from typing import TYPE_CHECKING, Any, Match, Optional, Pattern, TypeVar, Union
+from typing import TYPE_CHECKING, Any, List, Optional
 
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handlers._basehandler import BaseHandler
-from telegram.ext._utils.types import CCT, HandlerCallback
+from telegram.ext._handlers.basehandler import BaseHandler
+from telegram.ext._utils.types import CCT, RT, HandlerCallback
 
 if TYPE_CHECKING:
     from telegram.ext import Application
 
-RT = TypeVar("RT")
 
-
-class StringRegexHandler(BaseHandler[str, CCT]):
-    """Handler class to handle string updates based on a regex which checks the update content.
-
-    Read the documentation of the :mod:`re` module for more information. The :func:`re.match`
-    function is used to determine if an update should be handled by this handler.
+class StringCommandHandler(BaseHandler[str, CCT]):
+    """Handler class to handle string commands. Commands are string updates that start with
+    ``/``. The handler will add a :obj:`list` to the
+    :class:`CallbackContext` named :attr:`CallbackContext.args`. It will contain a list of strings,
+    which is the text following the command split on single whitespace characters.
 
     Note:
         This handler is not used to handle Telegram :class:`telegram.Update`, but strings manually
@@ -47,7 +44,7 @@ class StringRegexHandler(BaseHandler[str, CCT]):
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Args:
-        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`): The regex pattern.
+        command (:obj:`str`): The command this handler should listen for.
         callback (:term:`coroutine function`): The callback function for this handler. Will be
             called when :meth:`check_update` has determined that an update should be processed by
             this handler. Callback signature::
@@ -63,7 +60,7 @@ class StringRegexHandler(BaseHandler[str, CCT]):
             .. seealso:: :wiki:`Concurrency`
 
     Attributes:
-        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`): The regex pattern.
+        command (:obj:`str`): The command this handler should listen for.
         callback (:term:`coroutine function`): The callback function for this handler.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
@@ -71,33 +68,31 @@ class StringRegexHandler(BaseHandler[str, CCT]):
 
     """
 
-    __slots__ = ("pattern",)
+    __slots__ = ("command",)
 
     def __init__(
         self,
-        pattern: Union[str, Pattern[str]],
+        command: str,
         callback: HandlerCallback[str, CCT, RT],
         block: DVType[bool] = DEFAULT_TRUE,
     ):
         super().__init__(callback, block=block)
+        self.command: str = command
 
-        if isinstance(pattern, str):
-            pattern = re.compile(pattern)
-
-        self.pattern: Union[str, Pattern[str]] = pattern
-
-    def check_update(self, update: object) -> Optional[Match[str]]:
+    def check_update(self, update: object) -> Optional[List[str]]:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
             update (:obj:`object`): The incoming update.
 
         Returns:
-            :obj:`None` | :obj:`re.match`
+            List[:obj:`str`]: List containing the text command split on whitespace.
 
         """
-        if isinstance(update, str) and (match := re.match(self.pattern, update)):
-            return match
+        if isinstance(update, str) and update.startswith("/"):
+            args = update[1:].split(" ")
+            if args[0] == self.command:
+                return args[1:]
         return None
 
     def collect_additional_context(
@@ -105,10 +100,9 @@ class StringRegexHandler(BaseHandler[str, CCT]):
         context: CCT,
         update: str,  # skipcq: BAN-B301
         application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
-        check_result: Optional[Match[str]],
+        check_result: Optional[List[str]],
     ) -> None:
-        """Add the result of ``re.match(pattern, update)`` to :attr:`CallbackContext.matches` as
-        list with one element.
+        """Add text after the command to :attr:`CallbackContext.args` as list, split on single
+        whitespaces.
         """
-        if self.pattern and check_result:
-            context.matches = [check_result]
+        context.args = check_result
