+++ telegram/ext/_handlers/stringregexhandler.py
@@ -16,14 +16,14 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the MessageHandler class."""
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, TypeVar, Union
+"""This module contains the StringRegexHandler class."""
 
-from telegram import Update
+import re
+from typing import TYPE_CHECKING, Any, Match, Optional, Pattern, TypeVar, Union
+
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext import filters as filters_module
-from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._handlers.basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, HandlerCallback
 
 if TYPE_CHECKING:
@@ -32,22 +32,22 @@ if TYPE_CHECKING:
 RT = TypeVar("RT")
 
 
-class MessageHandler(BaseHandler[Update, CCT]):
-    """Handler class to handle Telegram messages. They might contain text, media or status
-    updates.
+class StringRegexHandler(BaseHandler[str, CCT]):
+    """Handler class to handle string updates based on a regex which checks the update content.
+
+    Read the documentation of the :mod:`re` module for more information. The :func:`re.match`
+    function is used to determine if an update should be handled by this handler.
+
+    Note:
+        This handler is not used to handle Telegram :class:`telegram.Update`, but strings manually
+        put in the queue. For example to send messages with the bot using command line or API.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Args:
-        filters (:class:`telegram.ext.filters.BaseFilter`): A filter inheriting from
-            :class:`telegram.ext.filters.BaseFilter`. Standard filters can be found in
-            :mod:`telegram.ext.filters`. Filters can be combined using bitwise
-            operators (& for and, | for or, ~ for not). Passing :obj:`None` is a shortcut
-            to passing :class:`telegram.ext.filters.ALL`.
-
-            .. seealso:: :wiki:`Advanced Filters <Extensions---Advanced-Filters>`
+        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`): The regex pattern.
         callback (:term:`coroutine function`): The callback function for this handler. Will be
             called when :meth:`check_update` has determined that an update should be processed by
             this handler. Callback signature::
@@ -63,8 +63,7 @@ class MessageHandler(BaseHandler[Update, CCT]):
             .. seealso:: :wiki:`Concurrency`
 
     Attributes:
-        filters (:class:`telegram.ext.filters.BaseFilter`): Only allow updates with these Filters.
-            See :mod:`telegram.ext.filters` for a full list of all available filters.
+        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`): The regex pattern.
         callback (:term:`coroutine function`): The callback function for this handler.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
@@ -72,40 +71,44 @@ class MessageHandler(BaseHandler[Update, CCT]):
 
     """
 
-    __slots__ = ("filters",)
+    __slots__ = ("pattern",)
 
     def __init__(
         self,
-        filters: Optional[filters_module.BaseFilter],
-        callback: HandlerCallback[Update, CCT, RT],
+        pattern: Union[str, Pattern[str]],
+        callback: HandlerCallback[str, CCT, RT],
         block: DVType[bool] = DEFAULT_TRUE,
     ):
         super().__init__(callback, block=block)
-        self.filters: filters_module.BaseFilter = (
-            filters if filters is not None else filters_module.ALL
-        )
 
-    def check_update(self, update: object) -> Optional[Union[bool, Dict[str, List[Any]]]]:
+        if isinstance(pattern, str):
+            pattern = re.compile(pattern)
+
+        self.pattern: Union[str, Pattern[str]] = pattern
+
+    def check_update(self, update: object) -> Optional[Match[str]]:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
-            update (:class:`telegram.Update` | :obj:`object`): Incoming update.
+            update (:obj:`object`): The incoming update.
 
         Returns:
-            :obj:`bool`
+            :obj:`None` | :obj:`re.match`
 
         """
-        if isinstance(update, Update):
-            return self.filters.check_update(update) or False
+        if isinstance(update, str) and (match := re.match(self.pattern, update)):
+            return match
         return None
 
     def collect_additional_context(
         self,
         context: CCT,
-        update: Update,  # skipcq: BAN-B301
+        update: str,  # skipcq: BAN-B301
         application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
-        check_result: Optional[Union[bool, Dict[str, object]]],
+        check_result: Optional[Match[str]],
     ) -> None:
-        """Adds possible output of data filters to the :class:`CallbackContext`."""
-        if isinstance(check_result, dict):
-            context.update(check_result)
+        """Add the result of ``re.match(pattern, update)`` to :attr:`CallbackContext.matches` as
+        list with one element.
+        """
+        if self.pattern and check_result:
+            context.matches = [check_result]
