+++ telegram/ext/_handlers/typehandler.py
@@ -16,31 +16,29 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the ChosenInlineResultHandler class."""
-import re
-from typing import TYPE_CHECKING, Any, Match, Optional, Pattern, TypeVar, Union, cast
+"""This module contains the TypeHandler class."""
 
-from telegram import Update
+from typing import Optional, Type, TypeVar
+
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._handlers.basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, HandlerCallback
 
 RT = TypeVar("RT")
-
-if TYPE_CHECKING:
-    from telegram.ext import Application
+UT = TypeVar("UT")
 
 
-class ChosenInlineResultHandler(BaseHandler[Update, CCT]):
-    """Handler class to handle Telegram updates that contain
-    :attr:`telegram.Update.chosen_inline_result`.
+class TypeHandler(BaseHandler[UT, CCT]):
+    """Handler class to handle updates of custom types.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Args:
+        type (:external:class:`type`): The :external:class:`type` of updates this handler should
+            process, as determined by :obj:`isinstance`
         callback (:term:`coroutine function`): The callback function for this handler. Will be
             called when :meth:`check_update` has determined that an update should be processed by
             this handler. Callback signature::
@@ -49,73 +47,49 @@ class ChosenInlineResultHandler(BaseHandler[Update, CCT]):
 
             The return value of the callback is usually ignored except for the special case of
             :class:`telegram.ext.ConversationHandler`.
+        strict (:obj:`bool`, optional): Use ``type`` instead of :obj:`isinstance`.
+            Default is :obj:`False`.
         block (:obj:`bool`, optional): Determines whether the return value of the callback should
             be awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`. Defaults to :obj:`True`.
 
             .. seealso:: :wiki:`Concurrency`
-        pattern (:obj:`str` | :func:`re.Pattern <re.compile>`, optional): Regex pattern. If not
-            :obj:`None`, :func:`re.match`
-            is used on :attr:`telegram.ChosenInlineResult.result_id` to determine if an update
-            should be handled by this handler. This is accessible in the callback as
-            :attr:`telegram.ext.CallbackContext.matches`.
 
-            .. versionadded:: 13.6
     Attributes:
+        type (:external:class:`type`): The :external:class:`type` of updates this handler should
+            process.
         callback (:term:`coroutine function`): The callback function for this handler.
+        strict (:obj:`bool`): Use :external:class:`type` instead of :obj:`isinstance`. Default is
+            :obj:`False`.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`.
-        pattern (`Pattern`): Optional. Regex pattern to test
-            :attr:`telegram.ChosenInlineResult.result_id` against.
-
-            .. versionadded:: 13.6
 
     """
 
-    __slots__ = ("pattern",)
+    __slots__ = ("type", "strict")
 
     def __init__(
         self,
-        callback: HandlerCallback[Update, CCT, RT],
+        type: Type[UT],  # pylint: disable=redefined-builtin
+        callback: HandlerCallback[UT, CCT, RT],
+        strict: bool = False,
         block: DVType[bool] = DEFAULT_TRUE,
-        pattern: Optional[Union[str, Pattern[str]]] = None,
     ):
         super().__init__(callback, block=block)
+        self.type: Type[UT] = type
+        self.strict: Optional[bool] = strict
 
-        if isinstance(pattern, str):
-            pattern = re.compile(pattern)
-
-        self.pattern: Optional[Union[str, Pattern[str]]] = pattern
-
-    def check_update(self, update: object) -> Optional[Union[bool, object]]:
+    def check_update(self, update: object) -> bool:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
-            update (:class:`telegram.Update` | :obj:`object`): Incoming update.
+            update (:obj:`object`): Incoming update.
 
         Returns:
-            :obj:`bool` | :obj:`re.match`
+            :obj:`bool`
 
         """
-        if isinstance(update, Update) and update.chosen_inline_result:
-            if self.pattern:
-                if match := re.match(self.pattern, update.chosen_inline_result.result_id):
-                    return match
-            else:
-                return True
-        return None
-
-    def collect_additional_context(
-        self,
-        context: CCT,
-        update: Update,  # skipcq: BAN-B301
-        application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
-        check_result: Union[bool, Match[str]],
-    ) -> None:
-        """This function adds the matched regex pattern result to
-        :attr:`telegram.ext.CallbackContext.matches`.
-        """
-        if self.pattern:
-            check_result = cast(Match, check_result)
-            context.matches = [check_result]
+        if not self.strict:
+            return isinstance(update, self.type)
+        return type(update) is self.type  # pylint: disable=unidiomatic-typecheck
