+++ tests/_files/test_inputmedia.py
@@ -607,7 +607,7 @@ class TestSendMediaGroupWithRequest:
         assert len(messages) == 3
         assert all(isinstance(mes, Message) for mes in messages)
         assert all(mes.media_group_id == messages[0].media_group_id for mes in messages)
-        assert all(mes.caption == f"photo {idx+1}" for idx, mes in enumerate(messages))
+        assert all(mes.caption == f"photo {idx + 1}" for idx, mes in enumerate(messages))
         assert all(
             mes.caption_entities == (MessageEntity(MessageEntity.BOLD, 0, 5),) for mes in messages
         )
@@ -742,7 +742,7 @@ class TestSendMediaGroupWithRequest:
             assert len(messages) == 3
             assert all(isinstance(mes, Message) for mes in messages)
             assert all(mes.media_group_id == messages[0].media_group_id for mes in messages)
-            assert all(mes.caption == f"photo {idx+1}" for idx, mes in enumerate(messages))
+            assert all(mes.caption == f"photo {idx + 1}" for idx, mes in enumerate(messages))
             assert all(
                 mes.caption_entities == (MessageEntity(MessageEntity.BOLD, 0, 5),)
                 for mes in messages
